Code improvements

+++ app/src/main/java/speedrun/example/com/speedrun/data/datasource/dto/TimesDto.kt
@@ -2,6 +2,6 @@ package speedrun.example.com.speedrun.data.datasource.dto
 
 data class TimesDto(
         val primary: String,
-        val primary_t: Int,
+        val primary_t: Long,
         val realTime: String?
 )
