+++ app/src/main/java/speedrun/example/com/speedrun/entity/TimesEntity.kt
@@ -2,6 +2,6 @@ package speedrun.example.com.speedrun.entity
 
 data class TimesEntity (
         val primary: String,
-        val primary_t: Int,
+        val primary_t: Long,
         val realTime: String?
 )
